refactor: move authentication model from `widgets` on `shared` layer
