+++ src/widgets/authorization/model/authorization.model.ts
@@ -1,12 +1,3 @@
-export interface IAuth {
-  email: string;
-  password: string;
-}
-
-export interface IFormData extends IAuth {
-  confirmPassword: string;
-}
-
 export interface IIconPassword {
   showPassword: boolean;
   setShowPassword: (value: boolean) => void;
